problem_test/hanlp_test/requirements_hanlp_env.txt
# 建议为 HanLP 句法分析单独建 venv 使用，避免与后端依赖冲突
#
# 说明：
# - HanLP 依存句法（CTB7_BIAFFINE_DEP_ZH）依赖 TensorFlow（hanlp[full] 会拉取）
# - TensorFlow 2.13.x 在 Windows 上对 typing-extensions 有严格上限（<4.6），
#   因此不能与后端（FastAPI/Pydantic2）装在同一个环境里

hanlp[full]

# 解析 docx（0.8.11 版本不强制要求较新的 typing-extensions）
python-docx==0.8.11

# 读取 .env / 连接 MySQL（可选，但建议装上）
python-dotenv
pymysql


